clean up domain check
